pyproject.toml
[project]
name = "project-twitter-follow-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "colorama>=0.4.6",
    "pyinstaller>=6.16.0",
    "pytest>=8.4.2",
    "python-decouple>=3.8",
    "selenium>=4.38.0",
    "undetected-chromedriver",
]

[tool.pyright]
typeCheckingMode = "basic"

[tool.uv.sources]
undetected-chromedriver = { git = "https://github.com/ultrafunkamsterdam/undetected-chromedriver/" }
